# Node/EX/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Aula Express API",
    "version": "1.0.0",
    "description": "Serviços feitos para aula de express"
  },
  "basePath": "/",
  "host": "localhost:8081",
  "schemes": ["http"],
  "tags": [
    {
      "name": "user",
      "description": "Operations about user"
    },
    {
      "name": "post",
      "description": "Operations about Post"
    }
  ],
  "paths": {
    "/users/{id}": {
      "get": {
        "tags": ["user"],
        "summary": "Find user by Id",
        "description": "Returns a single user",
        "operationId": "getUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Success operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/posts": {
      "post": {
        "tags": ["post"],
        "summary": "Create a single post",
        "operationId": "createPost",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Post",
            "in": "body",
            "description": "Post details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post created"
          },
          "400": {
            "description": "Invalid args"
          },
          "500": {
            "description": "Internal Error"
          }
        }
      },
      "get": {
        "tags": ["post"],
        "summary": "Get all posts",
        "operationId": "getAllPosts",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Returned all posts",
            "schema": {
              "$ref": "#/definitions/Posts"
            }
          },
          "500": {
            "description": "Internal Error"
          }
        }
      }
    },
    "/posts/{id}": {
      "put": {
        "tags": ["post"],
        "summary": "Edit a single post",
        "operationId": "editPost",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Post",
            "in": "body",
            "description": "Post details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post Edited "
          },
          "400": {
            "description": "Invalid args"
          },
          "500": {
            "description": "Internal Error"
          }
        }
      },
        "delete": {
          "tags": ["post"],
          "summary": "delete a single post",
          "operationId": "deletePost",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "Id",
              "in": "path",
              "description": "Post id",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Post Deleted "
            },
            "400": {
              "description": "Invalid args"
            },
            "500": {
              "description": "Internal Error"
            }
          }
        }
    }
  },
"definitions": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
             "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "sobrenome": {
            "type": "string"
          },
          "idade": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          }
        },
        "xml": {
          "name": "User"
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "titulo": {
            "type": "string"
          },
          "conteudo": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Post"
        }
      },
      "Posts": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Post"
        }
      }
  }
}